fix gui rendering

+++ common/src/main/java/com/lowdragmc/lowdraglib/gui/texture/GuiTextureGroup.java
@@ -52,7 +52,7 @@ public class GuiTextureGroup extends TransformTexture{
 
     @Environment(EnvType.CLIENT)
     @Override
-    protected void drawSubAreaInternal(PoseStack stack, float x, float y, int width, int height, float drawnU, float drawnV, float drawnWidth, float drawnHeight) {
+    protected void drawSubAreaInternal(PoseStack stack, float x, float y, float width, float height, float drawnU, float drawnV, float drawnWidth, float drawnHeight) {
         for (IGuiTexture texture : textures) {
             texture.drawSubArea(stack, x, y, width, height, drawnU, drawnV, drawnWidth, drawnHeight);
         }
